% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfunction.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{My First Function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{(variable) any number}
}
\value{
x+2
}
\description{
This function is my first function that adds 2 to any number.
}
